add daily quote hook and feature flag
