+++ components/AppContainer.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect } from 'react';
 import { Alert, SafeAreaView, StatusBar, StyleSheet } from 'react-native';
-import AsyncStorage from '@react-native-async-storage/async-storage';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { useTheme } from '@/contexts/ThemeContext';
-import usePushNotifications from '@/hooks/usePushNotifications';
 import { useAuth } from '@/contexts/AuthContext';
+import usePushNotifications from '@/hooks/usePushNotifications';
+import useDailyQuote from '@/hooks/useDailyQuote';
 
 export const AppContainer: React.FC<{ children: React.ReactNode }> = ({
   children,
@@ -17,6 +17,7 @@ export const AppContainer: React.FC<{ children: React.ReactNode }> = ({
       ? 'light-content'
       : 'dark-content';
   usePushNotifications();
+  useDailyQuote();
 
   useEffect(() => {
     if (authError) {
@@ -26,25 +27,9 @@ export const AppContainer: React.FC<{ children: React.ReactNode }> = ({
     }
   }, [authError, setAuthError]);
 
-  useEffect(() => {
-    const showQuote = async () => {
-      const enabled = await AsyncStorage.getItem('dailyQuote');
-      if (enabled === 'true') {
-        const quotes = [
-          'Believe in yourself!',
-          'Dream big and dare to fail.',
-          'Every day is a second chance.',
-        ];
-        const q = quotes[Math.floor(Math.random() * quotes.length)];
-        Alert.alert('Motivation', q);
-      }
-    };
-    showQuote();
-  }, []);
-
   return (
     <SafeAreaProvider>
-      <SafeAreaView style={[styles.container, { backgroundColor }]}>
+      <SafeAreaView style={[styles.container, { backgroundColor }]}> 
         <StatusBar barStyle={barStyle} />
         {children}
       </SafeAreaView>
