+++ constants/featureFlags.ts
@@ -1,0 +1,1 @@
+export const DAILY_QUOTE_ENABLED = true;
